requirements.txt
flask
langchain
langchain-groq
faiss-cpu
pypdf
python-dotenv
langchain-huggingface
sentence-transformers